% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-CTEM_FireMIP.R
\name{availableQuantities_CTEM_FireMIP}
\alias{availableQuantities_CTEM_FireMIP}
\title{List all quantities available for a FireMIP Source}
\usage{
availableQuantities_CTEM_FireMIP(source, names)
}
\arguments{
\item{source}{A path to a directory on the file system containing some .out files}
}
\value{
A list of all the .out files present, with the ".out" removed.
}
\description{
Simply lists all LPJ-GUESS output variables (stored as .out files) available in a directory.
Also ignores some common red herrings like "guess.out" and "*.out"
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
