% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-CLM_FireMIP.R
\docType{data}
\name{CLM_FireMIP}
\alias{CLM_FireMIP}
\title{CLM-FireMIP Format objects}
\format{
A \code{Quantity} object is an S4 class.
}
\usage{
CLM_FireMIP
}
\description{
\code{CLM_FireMIP} - a Format for reading CLM FireMIP model output
}
\keyword{datasets}
