% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-CTEM_FireMIP.R
\docType{data}
\name{CTEM_FireMIP}
\alias{CTEM_FireMIP}
\title{CTEM-FireMIP Format objects}
\format{
A \code{Quantity} object is an S4 class.
}
\usage{
CTEM_FireMIP
}
\description{
\code{CTEM_FireMIP} - a Format for reading CTEM FireMIP model output
}
\keyword{datasets}
