% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-Inferno_FireMIP.R
\name{determinePFTs_Inferno_FireMIP}
\alias{determinePFTs_Inferno_FireMIP}
\title{Return Inferno PFTs}
\usage{
determinePFTs_Inferno_FireMIP(x, variables)
}
\arguments{
\item{x}{A Source objects describing an  Inferno-FireMIP source}

\item{variables}{Currently ignored.}
}
\description{
Return Inferno PFTs
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
