% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-LPJ-GUESS-GlobFIRM_FireMIP.R
\name{determinePFTs_LPJ_GUESS_GlobFIRM_FireMIP}
\alias{determinePFTs_LPJ_GUESS_GlobFIRM_FireMIP}
\title{Detemine PFTs present in an FireMIP run source}
\usage{
determinePFTs_LPJ_GUESS_GlobFIRM_FireMIP(x, variables)
}
\arguments{
\item{x}{A Source objects describing a FireMIP source}

\item{variables}{Some variable to look for to detremine the PFTs present in the run.  Not the function automatically searches:
"lai", "cmass", "dens" and "fpc".  If they are not in your output you should define another per-PFT variable here.  Currently ignored.}
}
\description{
Detemine PFTs present in an FireMIP run source
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
