% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireMIP-PFTs.R
\docType{data}
\name{FireMIP.PFTs}
\alias{FireMIP.PFTs}
\title{FireMIP PFTs}
\format{An S4 class object with the slots as defined below.}
\usage{
FireMIP.PFTs
}
\description{
A coarse but simple PFT set to which all the FireMIP models PFTs may be classified.
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{datasets}
