% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-FireMIP.R
\docType{data}
\name{FireMIP}
\alias{FireMIP}
\title{FireMIP Format object}
\format{
A \code{Format} object is an S4 class.
}
\usage{
FireMIP
}
\description{
\code{FireMIP} - a Format for reading 2019 FireMIP model output
}
\keyword{datasets}
