% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-Inferno_FireMIP.R
\docType{data}
\name{Inferno_FireMIP}
\alias{Inferno_FireMIP}
\title{Inferno-FireMIP Format objects}
\format{
A \code{Quantity} object is an S4 class.
}
\usage{
Inferno_FireMIP
}
\description{
\code{Inferno_FireMIP} - a Format for reading Inferno FireMIP model output
}
\keyword{datasets}
