% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-JSBACH_FireMIP.R
\docType{data}
\name{JSBACH_FireMIP}
\alias{JSBACH_FireMIP}
\title{JSBACH-FireMIP Format objects}
\format{A \code{Quantity} object is an S4 class.}
\usage{
JSBACH_FireMIP
}
\description{
\code{JSBACH_FireMIP} - a Format for reading JSBACH FireMIP model output
}
\keyword{datasets}
