% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-LPJ-GUESS-GlobFIRM_FireMIP.R
\docType{data}
\name{LPJ_GUESS_GlobFIRM_FireMIP}
\alias{LPJ_GUESS_GlobFIRM_FireMIP}
\title{LPJ_GUESS_GlobFIRM-FireMIP Format objects}
\format{A \code{Quantity} object is an S4 class.}
\usage{
LPJ_GUESS_GlobFIRM_FireMIP
}
\description{
\code{LPJ_GUESS_GlobFIRM_FireMIP} - a Format for reading LPJ_GUESS_GlobFIRM FireMIP model output
}
\keyword{datasets}
