% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-LPJ-GUESS-SIMFIRE-BLAZE_FireMIP.R
\docType{data}
\name{LPJ_GUESS_SIMFIRE_BLAZE_FireMIP}
\alias{LPJ_GUESS_SIMFIRE_BLAZE_FireMIP}
\title{LPJ_GUESS_SIMFIRE_BLAZE-FireMIP Format objects}
\format{A \code{Quantity} object is an S4 class.}
\usage{
LPJ_GUESS_SIMFIRE_BLAZE_FireMIP
}
\description{
\code{LPJ_GUESS_SIMFIRE_BLAZE_FireMIP} - a Format for reading LPJ_GUESS_SIMFIRE_BLAZE FireMIP model output
}
\keyword{datasets}
