% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-LPJ-GUESS-SPITFIRE_FireMIP.R
\docType{data}
\name{LPJ_GUESS_SPITFIRE_FireMIP}
\alias{LPJ_GUESS_SPITFIRE_FireMIP}
\title{LPJ_GUESS_SPITFIRE-FireMIP Format objects}
\format{
A \code{Quantity} object is an S4 class.
}
\usage{
LPJ_GUESS_SPITFIRE_FireMIP
}
\description{
\code{LPJ_GUESS_SPITFIRE_FireMIP} - a Format for reading LPJ_GUESS_SPITFIRE FireMIP model output
}
\keyword{datasets}
