% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-MC2_FireMIP.R
\docType{data}
\name{MC2_FireMIP}
\alias{MC2_FireMIP}
\title{MC2-FireMIP Format objects}
\format{A \code{Quantity} object is an S4 class.}
\usage{
MC2_FireMIP
}
\description{
\code{MC2_FireMIP} - a Format for reading MC2 FireMIP model output
}
\keyword{datasets}
