% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-LPJ-GUESS-SPITFIRE_FireMIP.R
\name{openFireMIPOutputFile_LPJ_GUESS_SPITFIRE}
\alias{openFireMIPOutputFile_LPJ_GUESS_SPITFIRE}
\title{Open a LPJ_GUESS_SPITFIRE FireMIP output file}
\usage{
openFireMIPOutputFile_LPJ_GUESS_SPITFIRE(
  run,
  quantity,
  sta.info,
  file.name,
  verbose = TRUE
)
}
\arguments{
\item{run}{A Source object to define the run we want to open.}

\item{quantity}{A Quantity object to define which variable we want to look up}

\item{verbose}{Logical, if TRUE spew forth a lot of info.}

\item{first.year}{The first year we want to read (numeric)}

\item{last.year}{The last year we want to read (numeric)}

\item{spatial.extent}{The spatial extent we want to read (as defined by as raster::extent or an object that can be cast to a raster::extent)}
}
\value{
A list containaing a data.table and an STAInfo object
}
\description{
Opens a .nc file from the LPJ_GUESS_SPITFIRE FireMIP output and sorts out the meta-data and dimensions and all that messy stuff.
Returns a data.table, because it is intended to be called by getField(), but of course the data.table could be used directly if you wish
}
\author{
Matthew Forrest \email{matthew.forrest@senckenberg.de}
}
\keyword{internal}
