% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format-ORCHIDEE_FireMIP.R
\docType{data}
\name{ORCHIDEE_FireMIP}
\alias{ORCHIDEE_FireMIP}
\title{ORCHIDEE-FireMIP Format objects}
\format{
A \code{Quantity} object is an S4 class.
}
\usage{
ORCHIDEE_FireMIP
}
\description{
\code{ORCHIDEE_FireMIP} - a Format for reading ORCHIDEE FireMIP model output
}
\keyword{datasets}
