% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireMIP-Quantities.R
\docType{data}
\name{FireMIP.quantities}
\alias{FireMIP.quantities}
\title{FireMIp Quantities}
\format{The \code{Quantity} class is an S4 class with the slots defined below.}
\usage{
FireMIP.quantities
}
\description{
All the FiremMIP quantities currently provided by the models.
}
\keyword{datasets}
